pyproject.toml
[tool.poetry]
name = "power_plant_construction"
version = "0.1.0"
description = ""
authors = ["mustdieka"]

[tool.poetry.dependencies]
psycopg2-binary = "2.9.7"
gunicorn = "20.1.0"
orjson = "3.9.4"
python = "^3.11.4"
click = "^8.1.3"
alembic = "^1.7.7"
python-dotenv = "^0.20.0"
tenacity = "^8.0.1"
pydantic = "^1.9.1"
SQLAlchemy = "2.0.18"
asyncpg = "^0.27.0"
PyYAML = "^6.0"
fastapi = "^0.100.0"
uvicorn = "^0.17.6"
passlib = "^1.7.4"
nats-py = "^2.1.4"
bcrypt = "^3.2.2"
pyjwt = "^2.6.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^6.1"
mypy = "^0.950"
isort = "^5.10.1"
black = {extras = ["jupyter"], version = "^22.6.0"}
pylint = "^2.13.9"
flake8 = "^4.0.1"
safety = "^1.10.3"
ipython = "^8.3.0"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
freezegun = "^1.2.1"
pytest-asyncio = "^0.19.0"
licensecheck = "^2023.1.1"
asyncpg-stubs = "0.27.0"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0", "pip>=23.1.2", "setuptools>=68.0.0", "wheel>=0.40.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
from_first = true
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=110

[tool.black]
line-length = 110
target-version = ["py310"]

[tool.pylint.main]
extension-pkg-allow-list = ["pydantic"]
